requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
email-validator
sqlalchemy
psycopg2-binary
python-dotenv
gunicorn